requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
bcrypt==4.0.1
pytest==7.4.2
